+++ azure-eventhubs/azure/eventhub/aio/__init__.py
@@ -1,0 +1,9 @@
+from .event_hubs_client_async import EventHubClient
+from .receiver_async import Receiver
+from .sender_async import Sender
+
+__all__ = [
+    "EventHubClient",
+    "Receiver",
+    "Sender"
+]
