+++ azure-eventhubs/azure/eventhub/client.py
@@ -25,6 +25,7 @@ from azure.eventhub import __version__
 from azure.eventhub.sender import Sender
 from azure.eventhub.receiver import Receiver
 from azure.eventhub.common import EventHubError, parse_sas_token
+from .client_abstract import EventHubClientAbstract
 
 
 log = logging.getLogger(__name__)
@@ -85,7 +86,7 @@ def _build_uri(address, entity):
     return address
 
 
-class EventHubClient(object):
+class EventHubClient(EventHubClientAbstract):
     """
     The EventHubClient class defines a high level interface for sending
     events to and receiving events from the Azure Event Hubs service.
@@ -99,166 +100,6 @@ class EventHubClient(object):
             :caption: Create a new instance of the Event Hub client
 
     """
-
-    def __init__(
-            self, address, username=None, password=None, debug=False,
-            http_proxy=None, auth_timeout=60, sas_token=None):
-        """
-        Constructs a new EventHubClient with the given address URL.
-
-        :param address: The full URI string of the Event Hub. This can optionally
-         include URL-encoded access name and key.
-        :type address: str
-        :param username: The name of the shared access policy. This must be supplied
-         if not encoded into the address.
-        :type username: str
-        :param password: The shared access key. This must be supplied if not encoded
-         into the address.
-        :type password: str
-        :param debug: Whether to output network trace logs to the logger. Default
-         is `False`.
-        :type debug: bool
-        :param http_proxy: HTTP proxy settings. This must be a dictionary with the following
-         keys: 'proxy_hostname' (str value) and 'proxy_port' (int value).
-         Additionally the following keys may also be present: 'username', 'password'.
-        :type http_proxy: dict[str, Any]
-        :param auth_timeout: The time in seconds to wait for a token to be authorized by the service.
-         The default value is 60 seconds. If set to 0, no timeout will be enforced from the client.
-        :type auth_timeout: int
-        :param sas_token: A SAS token or function that returns a SAS token. If a function is supplied,
-         it will be used to retrieve subsequent tokens in the case of token expiry. The function should
-         take no arguments.
-        :type sas_token: str or callable
-        """
-        self.container_id = "eventhub.pysdk-" + str(uuid.uuid4())[:8]
-        self.sas_token = sas_token
-        self.address = urlparse(address)
-        self.eh_name = self.address.path.lstrip('/')
-        self.http_proxy = http_proxy
-        self.mgmt_target = "amqps://{}/{}".format(self.address.hostname, self.eh_name)
-        url_username = unquote_plus(self.address.username) if self.address.username else None
-        username = username or url_username
-        url_password = unquote_plus(self.address.password) if self.address.password else None
-        password = password or url_password
-        if (not username or not password) and not sas_token:
-            raise ValueError("Please supply either username and password, or a SAS token")
-        self.auth_uri = "sb://{}{}".format(self.address.hostname, self.address.path)
-        self._auth_config = {'username': username, 'password': password}
-        self.get_auth = functools.partial(self._create_auth)
-        self.debug = debug
-        self.auth_timeout = auth_timeout
-
-        self.clients = []
-        self.stopped = False
-        log.info("%r: Created the Event Hub client", self.container_id)
-
-    @classmethod
-    def from_sas_token(cls, address, sas_token, eventhub=None, **kwargs):
-        """Create an EventHubClient from an existing auth token or token generator.
-
-        :param address: The Event Hub address URL
-        :type address: str
-        :param sas_token: A SAS token or function that returns a SAS token. If a function is supplied,
-         it will be used to retrieve subsequent tokens in the case of token expiry. The function should
-         take no arguments.
-        :type sas_token: str or callable
-        :param eventhub: The name of the EventHub, if not already included in the address URL.
-        :type eventhub: str
-        :param debug: Whether to output network trace logs to the logger. Default
-         is `False`.
-        :type debug: bool
-        :param http_proxy: HTTP proxy settings. This must be a dictionary with the following
-         keys: 'proxy_hostname' (str value) and 'proxy_port' (int value).
-         Additionally the following keys may also be present: 'username', 'password'.
-        :type http_proxy: dict[str, Any]
-        :param auth_timeout: The time in seconds to wait for a token to be authorized by the service.
-         The default value is 60 seconds. If set to 0, no timeout will be enforced from the client.
-        :type auth_timeout: int
-
-        Example:
-            .. literalinclude:: ../examples/test_examples_eventhub.py
-                :start-after: [START create_eventhub_client_sas_token]
-                :end-before: [END create_eventhub_client_sas_token]
-                :language: python
-                :dedent: 4
-                :caption: Create an EventHubClient from an existing auth token or token generator.
-
-        """
-        address = _build_uri(address, eventhub)
-        return cls(address, sas_token=sas_token, **kwargs)
-
-    @classmethod
-    def from_connection_string(cls, conn_str, eventhub=None, **kwargs):
-        """Create an EventHubClient from a connection string.
-
-        :param conn_str: The connection string.
-        :type conn_str: str
-        :param eventhub: The name of the EventHub, if the EntityName is
-         not included in the connection string.
-        :type eventhub: str
-        :param debug: Whether to output network trace logs to the logger. Default
-         is `False`.
-        :type debug: bool
-        :param http_proxy: HTTP proxy settings. This must be a dictionary with the following
-         keys: 'proxy_hostname' (str value) and 'proxy_port' (int value).
-         Additionally the following keys may also be present: 'username', 'password'.
-        :type http_proxy: dict[str, Any]
-        :param auth_timeout: The time in seconds to wait for a token to be authorized by the service.
-         The default value is 60 seconds. If set to 0, no timeout will be enforced from the client.
-        :type auth_timeout: int
-
-        Example:
-            .. literalinclude:: ../examples/test_examples_eventhub.py
-                :start-after: [START create_eventhub_client_connstr]
-                :end-before: [END create_eventhub_client_connstr]
-                :language: python
-                :dedent: 4
-                :caption: Create an EventHubClient from a connection string.
-
-        """
-        address, policy, key, entity = _parse_conn_str(conn_str)
-        entity = eventhub or entity
-        address = _build_uri(address, entity)
-        return cls(address, username=policy, password=key, **kwargs)
-
-    @classmethod
-    def from_iothub_connection_string(cls, conn_str, **kwargs):
-        """
-        Create an EventHubClient from an IoTHub connection string.
-
-        :param conn_str: The connection string.
-        :type conn_str: str
-        :param debug: Whether to output network trace logs to the logger. Default
-         is `False`.
-        :type debug: bool
-        :param http_proxy: HTTP proxy settings. This must be a dictionary with the following
-         keys: 'proxy_hostname' (str value) and 'proxy_port' (int value).
-         Additionally the following keys may also be present: 'username', 'password'.
-        :type http_proxy: dict[str, Any]
-        :param auth_timeout: The time in seconds to wait for a token to be authorized by the service.
-         The default value is 60 seconds. If set to 0, no timeout will be enforced from the client.
-        :type auth_timeout: int
-
-        Example:
-            .. literalinclude:: ../examples/test_examples_eventhub.py
-                :start-after: [START create_eventhub_client_iot_connstr]
-                :end-before: [END create_eventhub_client_iot_connstr]
-                :language: python
-                :dedent: 4
-                :caption: Create an EventHubClient from an IoTHub connection string.
-
-        """
-        address, policy, key, _ = _parse_conn_str(conn_str)
-        hub_name = address.split('.')[0]
-        username = "{}@sas.root.{}".format(policy, hub_name)
-        password = _generate_sas_token(address, policy, key)
-        client = cls("amqps://" + address, username=username, password=password, **kwargs)
-        client._auth_config = {  # pylint: disable=protected-access
-            'iot_username': policy,
-            'iot_password': key,
-            'username': username,
-            'password': password}
-        return client
 
     def _create_auth(self, username=None, password=None):
         """
@@ -289,20 +130,6 @@ class EventHubClient(object):
                 self.address.hostname, username, password, http_proxy=self.http_proxy)
         return authentication.SASTokenAuth.from_shared_access_key(
             self.auth_uri, username, password, timeout=self.auth_timeout, http_proxy=self.http_proxy)
-
-    def create_properties(self):  # pylint: disable=no-self-use
-        """
-        Format the properties with which to instantiate the connection.
-        This acts like a user agent over HTTP.
-
-        :rtype: dict
-        """
-        properties = {}
-        properties["product"] = "eventhub.python"
-        properties["version"] = __version__
-        properties["framework"] = "Python {}.{}.{}".format(*sys.version_info[0:3])
-        properties["platform"] = sys.platform
-        return properties
 
     def _close_clients(self):
         """
