+++ azure-eventhubs/azure/eventprocessorhost/eh_partition_pump.py
@@ -5,7 +5,8 @@
 
 import logging
 import asyncio
-from azure.eventhub import Offset, EventHubClientAsync
+from azure.eventhub import Offset
+from azure.eventhub.aio import EventHubClient
 from azure.eventprocessorhost.partition_pump import PartitionPump
 
 
@@ -64,7 +65,7 @@ class EventHubPartitionPump(PartitionPump):
         """
         await self.partition_context.get_initial_offset_async()
         # Create event hub client and receive handler and set options
-        self.eh_client = EventHubClientAsync(
+        self.eh_client = EventHubClient(
             self.host.eh_config.client_address,
             debug=self.host.eph_options.debug_trace,
             http_proxy=self.host.eph_options.http_proxy)
